# docs/swagger/swagger.yaml
basePath: /api
definitions:
  APIKey:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key_id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  APIKeyResponse:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key_id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  APIKeyWithSecret:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key:
        type: string
      key_id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  APIKeysPage:
    properties:
      data:
        items:
          $ref: '#/definitions/APIKey'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
  Agent:
    properties:
      description:
        type: string
      id:
        type: string
      instructions:
        type: string
      llms:
        items:
          $ref: '#/definitions/AgentToLLM'
        type: array
      name:
        type: string
    type: object
  AgentResponse:
    properties:
      description:
        type: string
      id:
        type: string
      instructions:
        type: string
      llms:
        items:
          $ref: '#/definitions/AgentToLLM'
        type: array
      name:
        type: string
    type: object
  AgentToLLM:
    properties:
      llm:
        $ref: '#/definitions/LLM'
      llm_id:
        type: string
      priority:
        type: integer
    type: object
  AgentsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/Agent'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
...
  ClientAgentsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/Agent'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
...
  CreateAPIKeyDTO:
    properties:
      expires_at:
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
  CreateAgentDTO:
    properties:
      description:
        maxLength: 1000
        type: string
      instructions:
        maxLength: 1000
        type: string
      llm_ids:
        items:
          type: string
        minItems: 1
        type: array
      name:
        maxLength: 255
        type: string
    required:
    - llm_ids
    - name
    type: object
...
  Error:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  JSONMap:
    additionalProperties: {}
    type: object
...
  LLM:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
  RunAgentDTO:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  UpdateAgentDTO:
    properties:
      description:
        maxLength: 1000
        type: string
      instructions:
        maxLength: 1000
        type: string
      llm_ids:
        items:
          type: string
        minItems: 1
        type: array
      name:
        maxLength: 255
        type: string
    type: object
...
  UpdateExpirationDTO:
    properties:
      expires_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Snipet platform.
  title: Snipet API
  version: "1.0"
paths:
  /agent:
    get:
      description: Lists agents.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AgentsPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: List agents
      tags:
      - agent
    post:
      consumes:
      - application/json
      description: Creates a new agent.
      parameters:
      - description: Agent data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateAgentDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/AgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Create agent
      tags:
      - agent
  /agent/{id}:
    delete:
      description: Deletes an agent by ID.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Delete agent
      tags:
      - agent
    get:
      description: Returns an agent by ID.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AgentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Get agent
      tags:
      - agent
    put:
      consumes:
      - application/json
      description: Updates an agent by ID.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateAgentDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Update agent
      tags:
      - agent
  /agent/{id}/run:
    post:
      consumes:
      - application/json
      description: Runs an agent in playground mode, streaming the execution via SSE.
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Run input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/RunAgentDTO'
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Run agent
      tags:
      - agent
  /api-key:
    get:
      description: Lists API keys, with optional pagination.
      parameters:
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIKeysPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: List API keys
      tags:
      - api-key
    post:
      consumes:
      - application/json
      description: Creates a new API key. The plaintext key is only returned once,
        on creation.
      parameters:
      - description: API key data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateAPIKeyDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/APIKeyWithSecret'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Create API key
      tags:
      - api-key
  /api-key/{id}:
    delete:
      description: Deletes an API key by ID.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Delete API key
      tags:
      - api-key
    get:
      description: Returns an API key by ID.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIKeyResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Get API key
      tags:
      - api-key
  /api-key/{id}/active:
    put:
      description: Marks an API key as active.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Activate API key
      tags:
      - api-key
  /api-key/{id}/disabled:
    put:
      description: Marks an API key as disabled.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Disable API key
      tags:
      - api-key
  /api-key/{id}/expiration:
    put:
      consumes:
      - application/json
      description: Updates the expiration date of an API key.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      - description: Expiration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateExpirationDTO'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Update API key expiration
      tags:
      - api-key
  /api-key/{id}/roll:
    post:
      description: Rotates an API key, invalidating the previous secret. The new plaintext
        key is only returned once.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIKeyWithSecret'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Roll API key
      tags:
      - api-key
  /api-key/me:
    get:
      description: Returns the API key used to authenticate the current request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Get current API key
      tags:
      - api-key
...
  /clients/{code}/agents:
    get:
      description: Lists agents linked to a client, with optional pagination.
      parameters:
      - description: Client code
        in: path
        name: code
        required: true
        type: string
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ClientAgentsPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List client agents
      tags:
      - client
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"